ceph_snap_create
Just save the code as .py and execute on your ceph storage cluster. This code uses ceph api to invoke snapshots on any given pool which is defined on the ceph storage cluster. With this one could perform a snap restore on any openstack instances which is running on top of it.